fix: ai set runtime
