+++ lib/services/runtime/handlers/aiResponse.ts
@@ -1,15 +1,11 @@
 import { BaseNode } from '@voiceflow/base-types';
-import { replaceVariables, sanitizeVariables } from '@voiceflow/common';
 import { VoiceNode } from '@voiceflow/voice-types';
-import axios from 'axios';
 
-import Config from '@/config';
-import log from '@/logger';
 import { HandlerFactory } from '@/runtime';
 
-import AIAssist, { AIAssistLog } from '../../aiAssist';
 import { FrameType, Output } from '../types';
 import { addOutputTrace, getOutputTrace } from '../utils';
+import { fetchPrompt } from './utils/ai';
 import { generateOutput } from './utils/output';
 import { getVersionDefaultVoice } from './utils/version';
 
@@ -18,62 +14,7 @@ const AIResponseHandler: HandlerFactory<VoiceNode.AIResponse.Node> = () => ({
   handle: async (node, runtime, variables) => {
     const nextID = node.nextId ?? null;
 
-    if (!Config.ML_GATEWAY_ENDPOINT) {
-      log.error('ML_GATEWAY_ENDPOINT is not set, skipping generative node');
-      return nextID;
-    }
-
-    const ML_GATEWAY_ENDPOINT = Config.ML_GATEWAY_ENDPOINT.split('/api')[0];
-
-    const sanitizedVars = sanitizeVariables(variables.getState());
-    const prompt = replaceVariables(node.prompt, sanitizedVars);
-    const system = replaceVariables(node.system, sanitizedVars);
-
-    const { maxTokens, temperature, model, voice } = node;
-
-    let response: string | null = null;
-
-    if ((node as any).mode === 'Memory') {
-      const messages = [...(variables.get<AIAssistLog>(AIAssist.StorageKey) || [])];
-      if (system) messages.unshift({ role: 'system', content: system });
-
-      response = await axios
-        .post<{ result: string }>(`${ML_GATEWAY_ENDPOINT}/api/v1/generation/chat`, {
-          messages,
-          maxTokens,
-          temperature,
-          model,
-        })
-        .then(({ data: { result } }) => result)
-        .catch(() => null);
-    } else if ((node as any).mode === 'Prompt + Memory') {
-      const messages = [...(variables.get<AIAssistLog>(AIAssist.StorageKey) || [])];
-      if (system) messages.unshift({ role: 'system', content: system });
-      messages.push({ role: 'system', content: prompt });
-
-      response = await axios
-        .post<{ result: string }>(`${ML_GATEWAY_ENDPOINT}/api/v1/generation/chat`, {
-          messages,
-          maxTokens,
-          temperature,
-          model,
-        })
-        .then(({ data: { result } }) => result)
-        .catch(() => null);
-    } else {
-      if (!node.prompt) return nextID;
-
-      response = await axios
-        .post<{ result: string }>(`${ML_GATEWAY_ENDPOINT}/api/v1/generation/generative-response`, {
-          prompt,
-          maxTokens,
-          system,
-          temperature,
-          model,
-        })
-        .then(({ data: { result } }) => result)
-        .catch(() => null);
-    }
+    const response = await fetchPrompt(node as any, variables);
 
     if (!response) return nextID;
 
@@ -81,7 +22,7 @@ const AIResponseHandler: HandlerFactory<VoiceNode.AIResponse.Node> = () => ({
       response,
       runtime.project,
       // use default voice if voice doesn't exist
-      voice ?? getVersionDefaultVoice(runtime.version)
+      node.voice ?? getVersionDefaultVoice(runtime.version)
     );
 
     runtime.stack.top().storage.set<Output>(FrameType.OUTPUT, output);
